project restructure + profitability page sketch
