+++ lib/app/navigation/app_router.dart
@@ -4,6 +4,7 @@ import 'package:flutter/material.dart';
 import '../viewmodel/cost_calculator/form/cost_calculator_form.dart';
 import '../widget/page/cost_calculator_page.dart';
 import '../widget/page/default_page.dart';
+import '../widget/page/profitability_page.dart';
 
 part 'app_router.gr.dart';
