+++ lib/app/viewmodel/cost_calculator/form/cost_calculator_form.dart
@@ -10,7 +10,7 @@ import 'package:dfunc/dfunc.dart';
 
 import '../form_block.dart';
 import '../input.dart';
-import '../../../entity/product/product.dart';
+import '../../../../domain/entity/product/product.dart';
 
 part 'cost_calculator_form.g.dart';
