+++ lib/app/widget/cost_calculator/bottom_total_bar.dart
@@ -3,9 +3,9 @@ import 'package:flutter_mobx/flutter_mobx.dart';
 import 'package:get_it/get_it.dart';
 import 'package:logger/logger.dart';
 
-import '../../entity/product/product.dart';
+import '../../../domain/entity/product/product.dart';
+import '../../../domain/database/repo/product_repository.dart';
 import '../../viewmodel/cost_calculator/form/cost_calculator_form.dart';
-import '../../database/repo/product_repository.dart';
 import '../../util/space.dart';
 
 class BottomTotalBar extends StatelessWidget {
