+++ lib/domain/entity/block/block.dart
