+++ lib/domain/entity/block/block.freezed.dart
