+++ lib/domain/entity/parameter/parameter.dart
