+++ lib/domain/entity/parameter/parameter.freezed.dart
