+++ lib/domain/entity/product/product.freezed.dart
